+++ page-objects/ProductsPage.js
@@ -1,34 +1,28 @@
 import { expect } from "@playwright/test"
+import { Navigation } from "./Navigation.js"
 
 export class ProductsPage {
     constructor(page) {
         this.page = page;
         this.addButtons = page.locator('[data-qa="product-button"]')
-        this.basketCounter = page.locator('[data-qa="header-basket-count"]')
     }
  
     visit = async () => {
         await this.page.goto("/")
     }
 
-    getBasketCount = async () => {
-        await this.basketCounter.waitFor()
-        //return a number
-        const text = await this.basketCounter.innerText()
-        // "0" -> 0
-        const asNumber = parseInt(text, 10)
-        return asNumber
-        } 
+   
 
     addProductToBasket = async (index) => {
         const specificButton = this.addButtons.nth(index)
         
         await specificButton.waitFor()
         await expect (specificButton).toHaveText("Add to Basket")
-        const basketCountBeforeAdding = await this.getBasketCount()
+        const navigation = new Navigation(this.page)
+        const basketCountBeforeAdding = await navigation.getBasketCount()
         await specificButton.click()
         await expect (specificButton).toHaveText("Remove from Basket")
-        const basketCountAfterAdding = await this.getBasketCount()
+        const basketCountAfterAdding = await navigation.getBasketCount()
         expect (basketCountAfterAdding).toBeGreaterThan(basketCountBeforeAdding)
 
     }
